HackRF Scheduler prototype Fasta Selkov B
This prototype software examines the scheduler functions using a roadmap
similar to biological process.  It uses a membrane design label to
contain spectrum, allowing sample to fit within particular mapping, akin
to cellular process.
The design character hence is scaleable via "membrane" type as function
and should be very efficient.

What the observer should know..........

The author has successfully inserted unknown process via molecular
weight, executed the overlying function, and mapped process.  Here the
author commits to inserting bandwidth for molecular weight, as function,
and mapping process with interactive tools to design call out.

The assignment of spectrum to alphanumeric identifier is of particular
importance, in that the underlying code might be interactive with other
software prototype.  Hence a master list of the authors assignment is
undertaken.  The concept here is to assign coherent varaible and
function.
The useful nature of this exercise is that a master overlying frame of
sort and map exists, functionality should be assignablle , reproducable
and very very fast in that the overly is contained on high end
platforms.

It is of particular note that the language I assign should be given a
starting name, nomenclature, of which I should like to assign a name
given in honor to Eugene Selkov. Here was a superb master of the
sorting, bell shaped curve who freely gave of his talent..  That it
should cross boundries outside of the biological process here assigned
seems worthy and fitting.
Following fasta sorting, I propose "FastaSelkovB" as my topic name for
nomenclature process.

I think I would welcome input as to spectrum to membrane boundry, but
hope developers will notice the thought and care I have undertaken, in
the sense I give credit to the designers of the superb tools I routinely
adapt and use to good and productive ends.

The endpoint is that developers in this undertakeing would be able to
adapt at once all sorting process via FastA, by assignment, and
transposing bandwidth to the wide and freely availble fasta sorting
tools.  Very handy indeed.

This would enter the alphanumeric sort to 23 regions, by assignment, and
mirror FASTA by bell shaped curve function in alignment.  The authors
thought process identifies spectrum in the start position "FSB"....after
Fasta Selkov B", and demonstrates bandwidth assignment.